--- routes/api/project/location/load_balancer.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "../../../merged/project/location/load_balancer"
