--- routes/api/project/vm.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "../../merged/project/vm"
