+++ routes/merged/project/location/private_subnet.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class Clover
-  branch = lambda do |r|
+  hash_branch(:project_location_prefix, "private-subnet") do |r|
     r.get api? do
       private_subnet_list
     end
@@ -86,7 +86,4 @@ class Clover
       end
     end
   end
-
-  hash_branch(:api_project_location_prefix, "private-subnet", &branch)
-  hash_branch(:project_location_prefix, "private-subnet", &branch)
 end
