--- routes/web/project/firewall.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "../../merged/project/firewall"
